requirements.txt
# --- Core ---
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.1

# --- Generative AI APIs ---
openai>=1.30.1
anthropic>=0.24.0
elevenlabs>=0.2.27

# --- Visualización / pruebas opcionales ---
streamlit>=1.35.0

# --- Control de tokens, logs y utilidades ---
tiktoken>=0.6.0
loguru>=0.7.2

# --- Exploración notebooks ---
notebook>=7.1.0

streamlit>=1.32
requests>=2.31
pyyaml>=6.0

